fix index conflict
