+++ internal/datacoord/task_scheduler.go
@@ -29,6 +29,10 @@ import (
 	"github.com/milvus-io/milvus/pkg/log"
 )
 
+const (
+	reqTimeoutInterval = time.Second * 10
+)
+
 type taskScheduler struct {
 	sync.RWMutex
